notes_automation/requirements.txt
playwright==1.42.0
pytest==8.0.0
pytest-playwright==0.4.4
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.1
